Add charmpp@6.8.2/6.10.2 % gcc@10.2.0; install after gromacs@2020.4
charmpp@6.8.2 was originally deployed on Expanse only under the
mvapich2/2.3.4 distribution for use with NAMD v2.13. We will reconsider
deploying it and NAMD with other MPI distirbutions again in the future
after further performance testing is complete. The aim here at this
point is simply to include it in the standard application stack to be
supported moving forward.

Also note charmpp@6.10.2 enabled the +omp variant, while charmpp@6.8.2
used the default ~omp variant. It is unclear if this variant option has
any affect on the charm++ build and/or its use with NAMD. This may be a
suprious option in the charmpp Spack package. This should be reviewed
again in the future.

IMPORTANT: Both spec install scripts for charmpp cannot use the
SPACK_SPEC environment variable. The multi-valued variant options are
not parsed correctly by the spack spec and spack install commands for
some reason when included as an expanded environemnt variable. This has
been observed previously with other specs and remains an open issue.
This is why each complete spec is given explicitly on a single line.