Add dftbplus@19.1 % gcc@10.2.0 ^openmpi@4.0.5 to expanse/0.15.4/cpu
This spec will not concretize due to a 'multiple providers' error
associated with lapack [1]. This may be why it was only deployed on
Expanse using the intel/19.1.1.217 compiler and intel-mpi/2019.8.254.
This issue should be investigated further in the future and resolved.
We will test an Intel-based spec install once the Intel compiler and
Intel MPI dependency chain(s) have been added.

[1]

==> Error: Multiple providers found for 'lapack': ['intel-parallel-studio@professional.2020.2%gcc@10.2.0 cflags="-O2 -march=native" cxxflags="-O2 -march=native" fflags="-O2 -march=native" ~advisor~clck+daal~gdb~ilp64~inspector+ipp~itac+mkl+mpi~newdtags+rpath+shared+tbb~vtune auto_dispatch=none threads=none arch=linux-centos8-zen2', 'openblas@0.3.10%gcc@10.2.0 cflags="-O2 -march=native" cxxflags="-O2 -march=native" fflags="-O2 -march=native" ~consistent_fpcsr+ilp64+pic+shared threads=none arch=linux-centos8-zen2']