Add mpip@3.5 % gcc@10.2.0 ^openmpi@4.0.5
mpip@3.4.1 was initially deployed on Expanse. However, it failed to
build using the the default Spack package from the builtins repo for
Spack v0.15.4. The latest version of the package from the spack/spack
develop branch was downloaded and placed in the sdsc package repo. Note,
however, the new default variants had to be disabled to complete a
successful installation. There was a problem with the build locating the
libunwind header files [1] despite a copy of the library already
installed by spack. The library also did not appear in the
concretization of mpip. This should not affect the functionality of mpip
on Expanse as these disabled variants were previosuly not enabled by
default in the Spack package. However, this issue should be investigated
in the future and resolved.

[1]

...
checking libunwind.h usability... no
checking libunwind.h presence... no
checking for libunwind.h... no

  mpiP on x86_64 platforms requires libunwind.
  Please install libunwind and, if necessary, configure mpiP with
  appropriate CFLAGS and LDFLAGS settings.