Re-add custom intel compiler package to sdsc package repo
In order to generate a modulefile for the Intel compilers that is
consistent with the lmod hierarchical scheme, the version numbers in the
spack builtin package.py file must be adjusted. By default, the builtin
package.py file utilizes the Intel Parallel Studio version numbering,
which is not the same as the Intel compiler version(s) distributed in
that version of Intel Parallel Studio.

It's not yet clear to me why Spack has configured the versioning to be
based only on the Intel Parallel Studio version. This issue should be
revisited in the future and the question posed to the Spack developers.