Add custom orca package to sdsc package repo
The builtin spack package.py file for ORCA needs some work. There are a
number of issues with version/string matching the name of the tarballs
that need to be manually downloaded from the ORCA website. This
package.py file will likely need a complete rewrite in the future.

There may be an issue with using OpenMPI v3.1.4 with ORCA v4.2.1 on
Expanse. Try with OpenMPI v3.1.6 during the next rebuild.