Add gcc@8.4.0 as core compiler; start dependency chain with cuda@10.2.89
We attempted to build gcc@8.4.0 with +nvptx to enable GPU offloading via
OpenACC and OpenMP. However, the build failed. There are a number of
outstanding issues with the +nvptx spack build. See the gcc@8.4.0.sh
spec build script for a few links to related GitHub issues. We plan to
revisit this issue in the future and set +nvptx when the problem in
Spack has been resolved.

Also, please note the use of ldflags line with gcc@8.4.0 in the
compilers.yaml file and LD_LIBRARY_PATH modification for cuda in the
modules.yaml file. Both are set pointing to the location of NVIDIA driver
libraries deployed by Bright. These libraries are not included in the
Spack CUDA package install, but are required in some cases to build
CUDA-enabled packages.