Start openmpi@4.0.5 % gcc@8.4.0 dep chain for expanse/0.15.4/gpu
The version of OpenMPI was updated to 4.0.5 here from 4.0.4 used
in the current production software stack on Expanse due to a previously
encountered bug in 4.0.4, which was resloved by updating to 4.0.5.

Also, note here that fabrics=verbs. We may also want to consider using
the fabrics=verbs variant for all OpenMPI 4.x.x versions in
/expanse/0.15.4/cpu in the future to see if it resolves this on-going
issue [1], where users sometimes find their jobs hanging.

[1]

[exp-13-16][[20769,0],74][btl_tcp.c:559:mca_btl_tcp_recv_blocking] recv(41) failed: Connection reset by peer (104)