Add beast1@1.10.4/1.8.4, beast2@2.5.2/2.6.3 % gcc@8.4.0 to expanse/0.15.4/gpu
Note: The existing Spack package.py file for beast2 does not explicitly
have a dependency on libbeagle, which means there is not explicit CUDA
dependency. This issue should be revisited in the future. I suspect the
package.py file should be updated to reflect beast2's ability to utilize
beagle as well.