Start intel@19.1.2.254 package dependency chain for expanse/0.15.4/gpu
Note, however, there are compilation issues starting at OpenBLAS. The
problem may be related to this fix/patch [1] that was applied after the
release of Spack v0.15.4 and the intermediate commit our deployment is
currently based on. We will test this fix/patch in the next iterative
rebuild of the expanse/0.15.4/gpu instance still being developed.

[1]

https://github.com/spack/spack/pull/19273/files