Add newer cpio and intel-mkl packages to sdsc package repo
We ran into a problem with the older version of the cpio package not
installing successfully when attempting to install the intel-mkl package
with an intel compiler in the expanse/0.15.4/gpu Spack instance.
Updating the cpio and intel-mkl packages from the current spack/develop
branch resolved the issue. As such, these packages should likely be
removed from the sdsc package repository on a future update of the Spack
version prior to deploying the new production instances.