Replace intel@19.1.2.254 dep chain with intel@19.0.5.281 for expanse/0.15.4/gpu
We've found a number of issues compiling standard packages with the newer
version of the Intel compilers. The primary issue with most packages may
be related to the use of the -no-gcc compiler option in many of them.
Overriding the -no-gcc with -gcc-sys appears to work for most packages
when using the older Intel compilers. This issue should be eampied in
more detail in the future. There are a number of package changes / fixes
eliminating the -no-gcc flag in the current spack/develop branch. And
there appears to be a long history (~10 years) behind the use of the
-no-gcc flag with Intel compilers. The shortake seems to be that it
should no longer be used with most newer Intel compilers.