Remove -gcc-sys flag from intel@19.0.5.281 compiler.yaml config
Some packages -- e.g., FFTW -- require the -gcc-sys flag [1] to be set
for the build to succeed with the intel@19.0.5.281 compiler(s). We are
removing it from the general intel@19.0.5.281 compiler configuration in
the compiler.yaml file and will instead move setting the -gcc-sys at the
individual package-level within the spec build scripts themselves.

[1]

https://www.intel.com/content/www/us/en/develop/documentation/cpp-compiler-developer-guide-and-reference/top/compiler-reference/compiler-options/compiler-option-details/preprocessor-options/gcc-gcc-sys.html