Downgrade preferred version of gdbm from 1.21 to 1.19
During the first pass through building the packages of the new
expasne/0.17.1/cpu instance, we found a number of core packages still
preferred gdbm@1.19 as their default dependency, despite the builtin
spack v0.17.1 package for gdbm setting gdbm@1.21 as its preferred
default version. When several packages that relied on gdbm were to be
used as dependnecies for another packagem, if the gdbm version were
different, this would cause a conflict. Most notably, this occured with
python@3.8.12 using the preferred gdbm@1.21, while perl@5.34.0 defaulted
to using gdbm@1.19.